+++ macros/src/lib.rs
@@ -1,5 +1,6 @@
 //! Macros part of the stlog logging framework
 
+#![cfg_attr(feature = "spanned", feature(proc_macro_span))]
 #![deny(warnings)]
 
 extern crate proc_macro;
@@ -8,9 +9,12 @@ extern crate quote;
 #[macro_use]
 extern crate syn;
 
-use proc_macro::TokenStream;
+use proc_macro::{Span, TokenStream};
 
-use syn::ItemStatic;
+use syn::{spanned::Spanned, Error, ItemStatic};
+
+#[cfg(feature = "spanned")]
+mod spanned;
 
 /// An attribute to declare a global logger
 ///
@@ -20,16 +24,23 @@ use syn::ItemStatic;
 pub fn global_logger(args: TokenStream, input: TokenStream) -> TokenStream {
     let var = parse_macro_input!(input as ItemStatic);
 
-    assert_eq!(
-        args.to_string(),
-        "",
-        "`global_logger` attribute takes no arguments"
-    );
+    if !args.is_empty() {
+        return Error::new(
+            Span::call_site().into(),
+            "`global_logger` attribute takes no arguments",
+        )
+        .to_compile_error()
+        .into();
+    }
 
-    assert!(
-        var.mutability.is_none(),
-        "`#[global_logger]` can't be used on `static mut` variables"
-    );
+    if var.mutability.is_some() {
+        return Error::new(
+            var.span(),
+            "`#[global_logger]` can't be used on `static mut` variables",
+        )
+        .to_compile_error()
+        .into();
+    }
 
     let attrs = var.attrs;
     let vis = var.vis;
@@ -56,5 +67,36 @@ pub fn global_logger(args: TokenStream, input: TokenStream) -> TokenStream {
         pub static __STLOG_GLOBAL_LOGGER__: &(stlog::GlobalLog) = {
             &#ident
         };
-    ).into()
+    )
+    .into()
+}
+
+#[cfg(feature = "spanned")]
+#[proc_macro]
+pub fn error(input: TokenStream) -> TokenStream {
+    spanned::common(input, "error")
+}
+
+#[cfg(feature = "spanned")]
+#[proc_macro]
+pub fn warning(input: TokenStream) -> TokenStream {
+    spanned::common(input, "warning")
+}
+
+#[cfg(feature = "spanned")]
+#[proc_macro]
+pub fn info(input: TokenStream) -> TokenStream {
+    spanned::common(input, "info")
+}
+
+#[cfg(feature = "spanned")]
+#[proc_macro]
+pub fn debug(input: TokenStream) -> TokenStream {
+    spanned::common(input, "debug")
+}
+
+#[cfg(feature = "spanned")]
+#[proc_macro]
+pub fn trace(input: TokenStream) -> TokenStream {
+    spanned::common(input, "trace")
 }